properly supply subject class sids to policy enforcement
This is the analogue of `new-type-sids` and and `get-subject-types` in the existing
pipeline. We don't need it until the end of processing because we are not looking up
shacl shapes as we go.

I added special datatype inference for the @type predicate, which is what we currently
do. We may want to expand that in the future.