capture all generate-new-sid logic in hydrate-schema
When we generate subject ids we take into consideration a lot of additional criteria
besides whether an iri is used as a predicate. The end effect of using a pid in the old
stage pipeline was that it was included in the vocab refresh at the end of a
transaction. So, in order to achieve the same effect we need to consult the same
criteria when "hydrating" the vocab instead of refreshing from scratch.