initial dumb function to decompose iris into namespaces and names.
This function is dumb. It will not work long term. There are a ton of edge cases
that this function will miss. I'm committing it now because wanted to move on to
the meat of this task with a skeletal function in place. I will revisit and
reimplement this function, possibly with an IRI parser, if I can find a suitable
one.