add shape cache invalidation
This is a little bit faster and smarter than our old shape cache invalidation. The old
stage pipeline nukes the whole cache if it finds a subject with class NodeShape or
PropertyShape. This was a little error prone because most property shapes do not have
the class PropertyShape, so updates to those wouldn't get caught.

This is a bit more surgical - it only removes the shapes that were modified, and it
should detect all shape changes, assuming that all property shapes have an sh:path
property (which is in the spec) and that all node shapes have @type sh:NodeShape, which
we document they should.