switch to constant length subid vectors
This is to ensure a lexical ordering of subids in order for index range scans to
work. The structure of the new subid vectors is [<ns code>, <first 8 bytes of
name as a long>, <full iri>]. String comparison is slow, but it should rarely be
necessary because most iris will differ in either their ns or the first 8 bytes
of their names.