WIP Add error formatters in more places, update error-formatting fn to use the most general error
By default, malli will provide errors for every part of the syntax tree that has a failure. This can mean many errors for a large tree (such as our fql schema), with many repetitions (eg among all the leaves of an `or`), and a lot of lost context on how those errors are related (eg if they are `or`s, `and`s, or something else).

Rather than attempt to always consolidate those programmatically, this adds custom formatters to parts of the schema that can have good errors written about them, and uses the `-resolve-root-error` fn to retrieve them.