properly handle nil datatypes during insert
When I removed the `(some? <component)` when gate that was dropping flakes because of
nil datatypes, I needed to add a when-gate to the existing-dt lookup to stop things from
blowing up. In this pipeline we should never have a `nil` ::where/val for any component,
but ::where/datatype _can_ be nil.

Added some pr-str calls in the subid impl so you can see the nil value when it happens.