hydrate schema cache instead of rebuilding from scratch each stage
This removes the class pid generation problem - we don't know when something will be
used as a predicate in advance, so assigning a special subject id range won't work. So
instead we just check every new flake and if a subject is used as a predicate ever it
will go into the :schema cache.

At index writing (`storage/write-db-root`) time we pull out all the predicate sids and
store them as a vector in the index root. On load, we can reconstruct the whole schema
cache with just those pids.

This change is compatible with both the old data pipeline and the new one, as they both
utilize the :schema cache the same way. That's why it's safe to replace the
`vocab/vocab-map` call with `vocab/load-schema` in `storage/reify-db`.

This should be a big performance win at stage time for ledgers with big vocabularies, at
the expense of more storage space in the index root file.