fix vocab flake filter
In the `hydrate-schema` function we need to extract all of the predicate sids from a
sequence of new flakes. The old code used a `cond` wrongly presumed that the cases were
mutually exclusive, when in fact each condition always needs to be checked. This commit
removes the cond and instead always checks if the subject and the object of a flake
could be a predicate id.

Fixes #599