update shacl basic tests to use stage2
Most of these are very basic transformations, but there are a few tricks. sids are no
longer deterministically generated, so test results that depend on non-predefined sids
need to use a `str/starts-with` check instead of equality checks.

Datatype error messages are now shacl validation error messages instead of the advanced
validation exception because we no longer do advanced validation.