fix target-objects-of shapes targeting new objects
The old code built target-objects-of shapes that targeted the subject, but the new
flakes can have objects that are targeted by target-objects-of.

It's a bit unweildy but it works.

Also discarded the exception? tests because they for some reason hang in the repl if
they aren't exceptions and they are always accompanied by ex-message tests anyways.