preserve txn context for txn clause expansion
If a user supplies additional context for a transaction we need to preserve it after the
initial expansion for the second round when we expand the @type @json values.

The user cannot supply a null context, as that will clear the fluree context and we
won't be able to parse the output. Instead, if a user wants to clear the context they
can provide an empty context.

Unfortunately, while investigating I discovered that the `-subid` dbproto method calls
expand on the iris, which and always uses the default context. In order to get around
this, I had to introduce an expand? opt.