remove accidental shacl shape clobbering
Since the order the subject ids are created in is non-deterministic, sometimes a sid
will get an entry in the subj-mods map via the o-pred-shapes before the
s-pred-shapes. When that happened, the s-pred-shapes clobbered the metadata already in
the subj-mods map via the call to assoc:

```
;; this was bad
(-> subj-mods (assoc sid subject-meta))
```

In order to fix it I need to carefully update each entry of the subj-mod meta map to
avoid clobbering existing data.