Match otherwise empty id maps
Parses id maps of the form `{"@id" <some iri>}` into a dummy triple that doesn't
include any variables whose subject is the iri reference, object is the matched
"@id" predicate, and value is the iri with string datatype. This will serve as
an existence check on the iri itself as a subject during an index range scan.