Mark failing txn test as no longer pending (#646)
After rebasing on main the problem was fixed.

So reviewing this is low priority, and the best answer might be to close this stack of PRs unmerged (well, except the code formatting one).

I'm just not sure if the new test is really different enough in any important ways to warrant keeping it around. It looks a lot like another test in the same ns. I'm not sure why it was failing when the other one was passing. But they both pass now.