store sh:datatype pred constraints in index root pred vector
This allows datatype constraints to survive a db load. Instead of storing a map, which
would be sparse since not every predicate will have a datatype constraint, I'm storing
the predicate meta information as a tuple. Once we have more meta information that we
need to survive a load (like statistics for a query planner) we can refactor this to a
richer data structure - this is the smallest change we need right now.