update misc tests to use new tx semantics
It isn't possible to get the spo test working without concretizing the hash values,
either by trying to get a deterministic sort of the results or by just hardcoding the
hashes. Since the error result is much easier to interpret when comparing sets than the
results of `pred-match?` I opted to just hardcode the expected result set.

misc fixup