properly discard insert flakes when there is no solution match
If a triple component doesn't match a solution, then we need to discard the whole
triple. The code was making the mistake of assuming that if there was no sid in the
match, then there must be an iri. That is not the case, and this corrects that
assumption.

This also finishes updates on the credential tests. There's a bug with using `@base` and
`@vocab` in the context and this works around that by not using `@base` and `@context`,
which will be fixed in a followup PR.