remove obsolete bootrap code and add id flake at db creation
The `initial-tx` ledger creation opt has been broken for a while and is no longer the
correct way to initialize a ledger. So that has been removed. The only remaining
function in the `bootstrap` namespace was `blank-db`, which added the base flakes and
set the initial t value to -1.

When the initial t value is 0, none of the base flakes were findable via an index-range
query during the first transaction - when the db t was greater than or equal to the
flake t. This isn't a problem when the base flakes are just the identity flake, since we
don't need the identity flake for transacting, only for querying. After the initial
transaction the db t will be -1 and everything works just fine.