fix reversed predicates in where clause
Fixes https://github.com/fluree/core/issues/62

The problem was that in the call to `expand-keys` in `parse/parse-node-map` we were
throwing away the context of whether or not a predicate was an @reversed predicate.

It turns out we were expanding both subjects and objects again further down the call
stack, so I removed this top-level call to expand keys and added a call to the handy
`json-ld/details` function when calling `parse/parse-predicate`. The `details` function
returns the expanded iri plus a map of the context details, which helpfully contains the
`:reverse` key.

Then I altered the `where/->predicate` match creation function to add the
`::where/reverse` key so that bit of context survives after expansion.*

Then, when the patterns are created I reverse the order if a predicate has the
`::where/reverse` key set to true.

*The `where/->predicate` function had an arity that accepts a recur-n parameter that adds
the `::where/recur` key to the iri matcher, but it is never invoked anywhere.