update with-id to only expand id-key
Because we need to defer expansion until we can properly hold on to the @reverse
context, we can't just `expand-keys` and then call `with-id`. However, we do need to
expand in order to properly identify the id-key. It's a pickle!

I've updated `with-id` to return the given node with only the id-key expanded.