add author and txn to commit
Author priority is txn-signer, then ledger signer, then empty string.
Txn is the id of the txn iff there's only one stage per commit, otherwise we discard
it. This was an administrative decision that we don't care about provenance in that
scenario.

The new flakes in the commit required flake counts and hashes to be updated.