update nameservice lookup doc and calling convention
Confusingly, all the docs and arg names for the ns-proto/-lookup method indicated that
the ledger-alias was used to look up the head commit address. But in fact, the only
place where it is called uses the ledger head address. This updates the docs, arg names,
and calling convention to align with actual usage.