remove unused force? argument to -commit-update method
This argument wasn't used currently. If we plan on adding this functionality in
the future, then we should do it with two separate functions/methods instead of
tangling two different behaviors into the same function with a boolean argument.
Two functions will be easier to refactor and modify later, and will be much
easier to reason about.