serialize ns codes instead of nses; convert ns-code keys back to int
Our code for reading index files automatically converts the json map keys to
keywords, but neither namespace codes nor namespaces should be converted.
namespace codes should be integers, and namespaces should be strings. Instead of
changing our index reading code not to do the conversion, I opted to convert
then convert back for ns codes because it (should be) more efficient to convert
to keywords while parsing JSON instead of after the fact, and there will be much
more data that needs the conversion than will not. I'm not ecstatic about
converting integers to keywords and back to integers, but I think this is the
best approach given the fact that more data will need to be converted to
keywords than not, but we may want to revisit this later.