Use -txn-write method to write transactions
There's nothing in the connection contract that says it will have a store field,
so writing to a connection's storage directly breaks the connection abstraction.
We will (very soon) overhaul and possibly remove the connection protocols, but
there will still be a connection abstraction implemented through a connection
namespace with functions operating on connections to provide all the required
functionality.