remove options hash from storage protocol write method
The options hash is no longer used for any calls to the storage/write method. We
will most likely refine the storage protocol implementations in the near future
content address writes, but not using an options map

Options map arguments to protocol methods are problematic because it allows
implementations to leak implementation details outside of the protocol contract
because generic options maps are open ended. This has the potential of making
different protocol implementations not interchangeable because they take
different options keys changing their behavior, which defeats the purpose of
protocols.

Needing to specify arguments to protocol methods usually means at least one of 3
things: the protocol is at the wrong level of abstraction, too much
functionality is pushed into a single implementation and that the
implementations should be split (FileStore vs ContentAddressedFileStore, for
example), or too much functionality is pushed into a single method and the
methods should be split (storage/write vs storage/content-addressed-write, for
example).