refactor validation to use v-ctx for sh:closed
Some constraints need to be able to translate sids to iris, and then compact them, in
the result :message. The validate-constraint component already had a ton of args, most
of which are just passed on unaltered down the stack. I consolidated them into a context
map to handle this bag of stuff we need. This should also make it easier to add
additional data to the validation context if we need to in the future.

Also adds support for the sh:closed constraint and updates the test.