fix handling of targetObjectsOf target declarations
This required a bit of a refactor to make a clear distinction between focus nodes and
value nodes, in accordance with the spec. In all the other targets the focus node is
just the sid of the s-flakes that are passed in, if the the target matches it. In the
case of targetObjectsOf, though, the focus node may the objects of the focus nodes with
the targeted predicate.

When we were only returning flakes it was impossible to know if we should be targeting
the subject or the object. Now that distinction is clear.

In order to preserve this distinction I refactored all the constraint checks to accept
shape, focus-node, and value-nodes as explicit arguments, instead of stuffing them
inside the v-ctx map. I also found a ton of unused bindings and simpler ways to handle
validation result creation.