process shape flakes in reverse to preserve order
Because it builds up a node via `(update node <pred> conj <obj>)`, the `:i` index order
of the flake objects is reversed. This especially messes up SHACL property paths, which
are very order dependent. Processing the flakes in reverse order is a simple way to
maintain the proper object ordering.