rewrite sparql translator
This provides a much simpler architecture for adding additional SPARQL support.

A few external changes were necessary:
1) I changed the FQL schema for :order-by clauses to not strictly require lists, but any
sequential collection. This is fine, as the only place we use it just destructures the
values, we are in no way depending on list-specific behaviors.
2) The SPARQL translator tests for :bind patterns were requiring FQL that didn't
actually work. The syntax changed when we changed the where syntax, and the translated
:bind syntax was never executed.
3) :group-by clauses can contain multiple variables or expressions, and it was simpler
for this implementation to just always return a sequence, regardless if there was one or
many conditions.

I did modify the grammar a bit, only by factoring out the :Func term to make
:BuiltInCall processing simpler.

This is a WIP commit, I plan to clean it up in the next one after I address the final
failing test.