Accept rule sources passed in through the query opts
Accepts a rules graph with :reasoner-rules in the opts and a rules db
with :reasoner-rules-db. Both of these keys expect a list, but since
our reasoner only supports one rule source currently, the list should
only have one element. If both a graph and a db are provided, it will
use the db.