add support for multiple values per var in VALUES pattern
In SPARQL, a VALUES pattern can appear inside a WHERE clause alongside other
patterns. In FQL :values is its own clause. This mismatch makes translating directly
awkward, since we need to insert our translated pattern at a different point in the
parse tree than where we are at translation time.

We got around it by translating it into a :bind pattern. However, a :bind pattern can
only bind a single value to a variable, so we limited the scope of a VALUES pattern to
only a single value.

This gets around it by "tagging" the translated value pattern and then after the whole
parsing is done, moving it to the correct place in the tree: see `format-values`. It's
awkward, but it gives much more flexibility to the user.