Stop throwing error when namespace is missing.
Throwing errors in this case causes queries for iris with namespaces that
haven't previously been registered to throw an error while queries for iris that
have registered namespaces but don't exist in the database return an empty
result set. These two types of queries should behave the same because they are
both querying for non-existent iris.