correctly assign langString datatype to language tagged strings
We were not parsing language tagged strings to langString datatypes correctly in where
patterns or in values patterns, nor inferring a langString datatype in the presence of a
language tag.

I had to update the test expectation because now the language tagged string no longer
conforms to the `sh:maxCount 1` _and_ the `sh:datatype xsd:string` constraints, instead of
just the `sh:maxCount` constraint.