fix extra nesting in select * results
The SPARQL grammar only permits a single term in the select clause if that term is a
wildcard. If the wildcard is in a sequence (even if it's the only one) our select
pipeline will introduce an extra layer of nesting because it thinks it is working one
value at a time, instead of on multiple values.

In the course of investigating this I was able to refactor the WildcardSelector
implementation to be a bit simpler - the behavior hasn't changed at all.