fix handling of SPARQL UNION patterns
Infix notation is such a pain to deal with, but we can deal with it. The way the syntax
parse tree works has union patterns expressed as additional group patterns separated by
UNION.

This puts :GroupOrUnionGraphPattern handling in a strange place - it can return patterns
both in and out of the :union pattern. This necessitates special handling in the
:GraphPatternNotTriples term to properly unpack it.