preserve original commit times during migration
This introduces a private opt :time for jld-ledger/commit!. If you call commit! via the
iCommit protocol the :time opt will be dropped and generated, making it unavailable to
users of the public api.

The commit time is now generated in `enrich-commit-opts` if it is not supplied via
internal api.

I also restructured the enriched commit-opts to make it more clear what options are used
where.