make sure new genesis commit has same time as commit at t 1
When migrating a ledger that doesn't have a genesis commit, we create the genesis
commit. Unfortunately that hardcodes the genesis commit's time to the moment when the
migration was run, which would be newer than the rest of the commits on the ledger.

In order to maintain consistent commit times, this adds the ability to backdate the
genesis commit to the time of the original first commit. It is done via an internal-only
opt to jld-ledger/create*. If the opt is passed in via the public api it will be removed
by `parse-ledger-options`.