remove IRI record types and use TypedValues instead
We need the ability to get non-inferable datatypes from expression evaluation. We were
getting around it in the case of iris by returning a typed value - the IRI record. A
more general approach that won't require as much work is returning a TypedValue from
expressions that need to provide datatypes as well.