use TypedValues for all function arguments and return types
In the `compile` phase we transform every literal into a TypedValue and convert any
variable matches into TypedValues. This allows us to always keep a value's datatype
along with it, which is crucial for non-inferrable datatypes (like "xsd:dateTime" or
"@id"). Everything needed to change so that we can transparently compose all the
functions together without adding checks of whether the incoming arguments are typed
values or not.