revamp time handling
After extensive investigation I've revamped how we're handling time. It's worth
retelling the story so we don't have to repeat the journey.

We coerce ISO8601 strings into LocalDateTimes or OffsetDateTimes. In order to compare
them I iniitally coerced them to Instants, using UTC as the offset for LocalDateTimes,
because our `now` function returned an Instant.

Unfortunately, Clojure pretty prints Instants as a tagged literal, like this:

> #inst "<timestamp>"

The Clojure reader, on the other hand, reads those tagged literals as java.util.Date
values. So when it came time to compare I had to juggle LocalDateTimes, OffsetDateTimes,
Instants, and java.util.Dates, with a whole lot of ugly converting back and forth.

An OffsetDateTime is basically just an Instant with a broader api, so by changing the
`now` function to return odts at UTC we get everything an Instant gives us, minus the
tagged literal nonsense.

We now only support LocalDateTimes and OffsetDateTimes, which keeps everything much
simpler.

For comparisons we now convert LocalDateTimes to OffsetDateTimes at UTC.
For extracting components from a datetime we convert to LocalDateTime just for
convenience and to avoid reflection.