refactor subquery handling to respect semantics of parent pattern
We were handling subqueries in isolation correctly by executing them beforehand, but
when they are placed in the scope of another where pattern those semantics were not
being respected.

This refactors our subquery support to process them along with the other types of where
patterns in a search. It closes over the subquery and the query executor so we can
initiate an entire query from within the search processing pipeline.