coerce literals to their actual values for comparison
If you provide a date literal {"@value" "2024-01-01" "@type" "xsd:date"} we need to
convert the string to an actual LocalDate object before we compare.

`coerce` used to only dispatch on datatype sid, but we don't have sids for datatype
literals and I didn't want to thread db into the execution chain so we could generate
them. So I made `coerce` work for datatype iris as well.