support multiple shacl targets and larger target shapes
When we build shapes with `build-shape-node`, we do it greedily - we keep recursing
until we find the "leaves" of the graph. The code assumed that a constructed shape would
have a structure like this:

```
{"@id" "ex:MyShape"
 "sh:targetClass" ["ex:SomeClass"]}
```

But because of greedy construction, it very well could have a shape like this:
```
{"@id" "ex:MyShape"
 "sh:targetClass" [{"@id" "ex:SomeClass" "ex:foo" "some-other-data"}]}
```

Also, most of the targets assumed that there would only be a single target, when
according to the spec there could be multiple.

This commit regularizes the shape structure to handle multiple targets and also
additional data on the target subject.

fixes https://github.com/fluree/core/issues/132