fix address split on non-jvm platforms
For some reason the regex #":(?!.*:)" was producing a different result on the browser
and in nodejs than on the jvm:

; clj:
(split-address "fluree:memory://testledger")
["fluree:memory" "//testledger"]

; cljs:
(split-address "fluree:memory://testledger")
["fluree" "memory://testledger"]

All the documentation I could find indicates that this regex should in fact work
cross-platform, but for some reason it reliably did not. I have replaced its use with an
alternate implementation that safely avoids the issue.