move publish-commit back to connection
ledger/create is just a thin wrapper around ledger/instantiate, so in order to preserve
ledger's ignorance of nameservices and connection internals I moved the basic setup to
the connection. This has the nice side effect of removing the need for the internal-only
opt for instantiating init-time during migration, since the call site is now responsible
for declaring the init time.