allow targetObjectsOf to target literal values
Our implementation of targetObjectsOf targets presupposed that the objects in question
would be subject references. This leaves a lot of utility for that target type on the
table.

In order to enable it I needed to make a "focus-node" and a "value-node" structurally
identical - in this case, a tuple of [value dt lang]. Then, in `validate-node-shape` in
the case that the focus node is not the a member of the s-flakes nor a reference, it
becomes the value node for the constraint.

As a note for future work, this is functionally equivalent to a `where/TypedValue`
record, we should consolidate those abstractions.