fix neverending loop while turning a long into bytes
We encountered a OOM error while trying to turn this IRI into a SID:
http://dbpedia.org/resource/Permian–Triassic_extinction_event

(note the unicode em-dash instead of a hyphen)

This reimplementation avoids an infinite loop. It also preserves a bug in the original
impl wherein zero bytes are removed, regardless of whether they were padding or not. A
fix for that will appear in a subsequent PR.