use OR semantics to compose policies
Before, if we hada bunch of policies we segregated them into different performance
tiers:
1. property polices (fastest)
2. class policies
3. default policies

The semantics were to only check the fastest policies for access - if you had a class
policy that granted access to a flake, but a property policy denied access, then you
would not get access.

The new semantic is that we will try each policy (in order of efficiency), and if a
policy does not grant access we will check the next policy. Once a single policy grants
access, we stop checking, and if we get through all the policies and nothing grants
access, then we deny access.