remove "duplicate" flakes when scanning an index
There are two reasons a flake can be "stale":
1. The flake has been retracted
2. An identical flake was added in a later transaction

When the `remove-stale-flakes` function was refactored for performance, only stale
flakes in the first sense were removed.

In the second sense, we check if flakes are semantically identical by comparing the
values: s, p, o, dt, m. However, our sorted sets also take into consideration `t`, so
two flakes with same semantic content but with different t values can coexist in our
indexes.

This change adds back support for the second sense of "stale" and attempts to preserve
as much of the performance boost as possible by reusing the transient flakes set.