do not run index scan with non-comparable s or p values
An index scan with a subject or predicate component of a SID or a `nil` value (match
everything) or may return data, but a literal value will not - all subjects and
predicates are SIDs in the indexes and cannot be compared to non-SID values.

It is entirely legal to bind a literal object value to a variable and then use that
variable in the subject or predicate position. When we detect a non-comparable s or p
value we can just skip the index scan.

This fixes a bug wherein we would try to get the sid from the s-mch and p-mch
components and incorrectly compute that as `nil`, which matches everything, leading to
erroneous query results.