allow users to specify ledger-specific query opts
In a dataset, the opts provided with the query (or the override opts) will be applied
for each ledger in the dataset. Sometimes this is not desirable, for example when you
want a given policy class to be only used for a subset of the ledgers in a dataset.

This commit introduces ledger-specific opts.

:opts {<ledger-alias> {...(ledger-specific-opts)}}

Right now the only ledger specific opts supported are :policy, :policy-class, and
:policy-values. Every other opt applies to the query as a whole (:max-fuel, :meta,
:format, :output, :identity). If a value is specified both under a ledger alias and at
the top level of an opts map, then the ledger-specific value will override the top-level
one.