only use shacl flakes when rebuilding shapes
The build-shape-node function was overly eager, following every ref flake and
constructing a node out of it.

For example, in this shape:
```
{"@id"   "ex:shape1",
 "@type" "sh:NodeShape",
 "sh:targetClass" {"@id" "ex:User"},
 "sh:property" [{"sh:path" {"@id" "ex:name"},
                 "sh:maxCount" 1}]}
```
When rebuilding, would produce something like this:
```
{#SID[1 "id"] #SID[101 "shape1]
 #SID[3 "type"] [#SID[5 "NodeShape"]]
 #SID[5 "targetClass"] [{#SID[1 "id"] #SID[101 "User"]
                         #SID[3 "type"] [#SID[106 "Class"]]
                         ,,, <any other flakes associated with ex:User>,,,}]
 #SID[5 "property"]
   [{#SID[1 "id"] #SID[24 "fdb1"]
     #SID[5 "path"] [{#SID[1 "id"] #SID[101 "name"]
                      #SID[3 "type"] [#SID[107 "Property"]]
                      ,,,<any other properties associated with ex:name>,,,}]
     #SID[5 "maxCount"] [1]}]}
```
And so on recursively until the hardcoded and arbitrary depth limit of 10.

Now we only recursively build shacl ref flakes, so we aren't pulling in any extraneous
data (and charging the user fuel for it). I don't believe we need the depth limit any
more, but I'm retaining for now as a safety.