manage service joins purely locally
This is much more inefficient, but some remote services can't handle a VALUES clause
with variables that aren't in the query. We'll need to adding SERVICE var filtering in
the future anyways to account for $-prefixed vars in the SERVICE clause, but this is the
simple and correct way to handle it for now.