remove VALUES clause functions
These will be a useful performance enhancement, but aren't needed for this first
iteration. The only additional validation they require is to only bind vars that are
present in the SERVICE clause - I ran into some errors from 3rd party services when
trying to handle vars not in the clause (though that is not strictly in spec for
SPARQL).